requirements.txt
fastapi==0.128.0
uvicorn==0.40.0

sqlalchemy==2.0.45
psycopg2-binary==2.9.11

pydantic==2.12.5
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-dotenv==1.2.1

langdetect==1.0.9

google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
google-genai==1.61.0

httpx==0.28.1
requests==2.32.5

numpy==2.2.6
scikit-learn==1.7.2
scipy==1.15.3
pandas==2.3.3

tqdm==4.67.1